requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.2
gunicorn==21.2.0
h5py==3.9.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
keras==2.13.1
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
Pillow==10.0.0
protobuf==4.23.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.31.0
requests-oauthlib==1.3.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.2
six==1.16.0
SQLAlchemy==2.0.20
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-cpu==2.13.1
tensorflow-estimator==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing-extensions==4.5.0
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
xgboost==1.7.6
